Cambios en ProductoDAO e IniciarSesion
Ahora se añade el DTO de Usuario como atributo a la sesión usando la api rest de Usuario. También se cambio el argumento de consultar productos para que el id sea "long" y no "int".

+++ ecommercePersistencia/src/main/java/IPersistencia/IProductoDAO.java
@@ -10,9 +10,9 @@ import java.util.List;
 public interface IProductoDAO {
     public List<Producto> consultarProductos();
     
-    public Producto consultarProducto(int idProducto);
+    public Producto consultarProducto(long idProducto);
     
-    public void elimianarProducto(int idProducto);
+    public void elimianarProducto(long idProducto);
     
     public void crearProducto(Producto producto);
